+++ packages/arch_logger/lib/blocpod_arch_logger.dart
@@ -3,3 +3,4 @@ library;
 
 export 'src/blocpod_event_logger.dart';
 export 'src/event_log_record_formatter.dart';
+export 'src/pretty_event_log_record_formatter.dart';